+++ src/app/modules/student-answers/models/text.ts
@@ -9,11 +9,18 @@ export enum TextStatus {
 	Unknown = 'unknown'
 }
 
+export enum ProcessingStatus {
+	NotProcessed = 'not-proc',
+	Processing = 'processing',
+	Processed = 'processed'
+}
+
 export interface Text {
 	uuid: string;
 	lang: SupportedLanguage;
 	raw: string;
 	sent?: string;
 	processed?: string;
+	processingStatus?: ProcessingStatus;
 	status: TextStatus;
 }