+++ src/app/state/global.state.ts
@@ -1,0 +1,5 @@
+export enum ResourceStatus {
+	NotLoaded = 'not-loaded',
+	Loading = 'loading',
+	Loaded = 'loaded'
+}