+++ src/environments/environment.prod.ts
@@ -2,5 +2,8 @@ export const environment = {
 	production: true,
 	datasetManager: {
 		url: 'https://dataset-manager.jlarteaga.dev'
+	},
+	coordinator: {
+		url: 'https://nlp-coordinator.jlarteaga.dev'
 	}
 };