+++ CHANGELOG.md
@@ -17,9 +17,16 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Fixed
 
+## [0.2.0]
+
+### Added
+
+- Added single question view.
+- Added freeling display component.
+
 ## [0.1.3]
 
-### Fixed student-answer translation component showing the rawStatement of the question instead of its sentStatement. 
+### Fixed student-answer translation component showing the rawStatement of the question instead of its sentStatement.
 
 ## [0.1.2]
 
@@ -50,7 +57,9 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Added component to translate student answers.
 - Added small button to sanitize some common issues.
 
-[Unreleased]: https://github.com/jlarteaga/thesis-dataset-editor/compare/0.1.2...develop
+[Unreleased]: https://github.com/jlarteaga/thesis-dataset-editor/compare/0.2.0...develop
+
+[0.2.0]: https://github.com/jlarteaga/thesis-dataset-editor/compare/0.1.2...0.2.0
 
 [0.1.2]: https://github.com/jlarteaga/thesis-dataset-editor/compare/0.1.1...0.1.2